read_channel/requirements.txt
bottle==0.12.19
gevent==21.1.2
greenlet==1.1.0
kafka-python==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
